+++ crates/consensus/src/transaction/recovered.rs
@@ -6,12 +6,14 @@ use derive_more::{AsRef, Deref};
 /// Signed object with recovered signer.
 #[derive(Debug, Clone, Copy, PartialEq, Hash, Eq, AsRef, Deref)]
 #[cfg_attr(feature = "arbitrary", derive(arbitrary::Arbitrary))]
+#[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
 pub struct Recovered<T> {
     /// Signer of the type
     signer: Address,
     /// Signed object
     #[deref]
     #[as_ref]
+    #[cfg_attr(feature = "serde", serde(flatten))]
     inner: T,
 }