+++ crates/eips/src/eip7685.rs
@@ -129,6 +129,7 @@ impl Requests {
 /// needed to simulate the presence of requests without holding actual data.
 #[derive(Debug, Clone, PartialEq, Eq, Hash, derive_more::From)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "serde", serde(untagged))]
 pub enum RequestsOrHash {
     /// Stores a list of requests, allowing for dynamic requests hash calculation.
     Requests(Requests),
@@ -162,6 +163,16 @@ impl RequestsOrHash {
             Self::Requests(requests) => Some(requests),
             Self::Hash(_) => None,
         }
+    }
+
+    /// Returns `true` if the variant is a list of requests.
+    pub const fn is_requests(&self) -> bool {
+        matches!(self, Self::Requests(_))
+    }
+
+    /// Returns `true` if the variant is a precomputed hash.
+    pub const fn is_hash(&self) -> bool {
+        matches!(self, Self::Hash(_))
     }
 }