+++ crates/provider/src/layers/chain.rs
@@ -1,8 +1,7 @@
+use crate::{Provider, ProviderLayer};
 use alloy_chains::NamedChain;
 use alloy_network::Network;
 use std::time::Duration;
-
-use crate::{Provider, ProviderLayer};
 
 /// A layer that wraps a [`NamedChain`]. The layer will be used to set
 /// the client's poll interval based on the average block time for this chain.
@@ -12,8 +11,13 @@ use crate::{Provider, ProviderLayer};
 pub struct ChainLayer(NamedChain);
 
 impl ChainLayer {
+    /// Create a new `ChainLayer` from the given chain.
+    pub fn new(chain: NamedChain) -> Self {
+        Self(chain)
+    }
+
     /// Get the chain's average blocktime, if applicable.
-    pub const fn average_blocktime_hint(&self) -> Option<Duration> {
+    pub fn average_blocktime_hint(&self) -> Option<Duration> {
         self.0.average_blocktime_hint()
     }
 }