+++ crates/rpc-types-eth/src/account.rs
@@ -63,6 +63,18 @@ pub struct EIP1186AccountProofResponse {
     pub storage_proof: Vec<EIP1186StorageProof>,
 }
 
+#[cfg(feature = "serde")]
+impl EIP1186AccountProofResponse {
+    /// After `SpuriousDragon` empty account is defined as account with nonce == 0 && balance == 0
+    /// && bytecode = None (or hash is [`KECCAK_EMPTY`](alloy_consensus::constants::KECCAK_EMPTY)).
+    pub fn is_empty(&self) -> bool {
+        self.nonce == 0
+            && self.balance.is_zero()
+            && self.storage_hash.is_zero()
+            && self.code_hash == alloy_consensus::constants::KECCAK_EMPTY
+    }
+}
+
 /// Extended account information (used by `parity_allAccountInfo`).
 #[derive(Clone, Debug, Default, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]