+++ crates/serde/src/other/arbitrary_.rs
@@ -13,7 +13,6 @@ impl arbitrary::Arbitrary<'_> for OtherFields {
 
 /// Redefinition of `serde_json::Value` for the purpose of implementing `Arbitrary`.
 #[derive(Clone, Debug, arbitrary::Arbitrary)]
-#[allow(unnameable_types)]
 enum ArbitraryValue {
     Null,
     Bool(bool),